Conexion Develop (#20)
Correcion de Conexion (Servidor) y Logica (Cliente) en la rama Develop

+++ src/AbstractFactory/PublicaFactory.java
@@ -6,7 +6,7 @@ import Modelo.Sala;
 
 public class PublicaFactory implements ClinicaFactory {
     @Override
-    public Medico crearMedico(int id, String nombre, String especialidad) {
+    public Medico crearMedico(Long id, String nombre, String especialidad) {
         return new Medico(id, nombre, especialidad, "PUBLICO");
     }
