+++ src/DAO/FacturaDAO.java
@@ -1,11 +1,12 @@
 package DAO;
 
 import Modelo.Factura;
+
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
 
 public interface FacturaDAO {
     CompletableFuture<List<Factura>> listarTodos();
     CompletableFuture<Factura> guardar(Factura factura);
-    CompletableFuture<Void> eliminar(int id);
+    CompletableFuture<Void> eliminar(long id);
 }
